sending username to server for player creation, creating and updating clientside graphics

+++ src/common/schemas/objectSchema.ts
@@ -3,8 +3,8 @@ import { Binary, defineSchema, Schema } from 'nengi';
 export const objectSchema: Schema = defineSchema({
   x: Binary.Float32,
   y: Binary.Float32,
-  width: Binary.Float32,
-  height: Binary.Float32,
+  width: Binary.UInt8,
+  height: Binary.UInt8,
   color: Binary.Float64,
   shape: Binary.String,
 });
